Remove ServerError exception
According to the specification each invalid response should return an json
object ("error") which will provide information what happen on the server.
There is no reason why jimson should take care of that and translate it to
something else. User should expect to receive for each request (no matter if
this response will success or fail, except Notification requests) an json object
with details about status of that response.